bug fix : non-compressed key of the v2p cache
Previously, LibVMI compressed the key of the v2p cache from the combination of vaddr (64bit) and dtb (64bit) into a 64bit result. 
Although it's usually efficient and safe with 32bit guest virtual address space, it may cause cache inconsistency when we operate a x86-64 guest.
For example, there is a debug info in my test over a Ubuntu 64bit guest.
--V2P cache hit 0x0000880000400000 -- 0x0000000033e56000 (0x000000003b40a000/0x0000000000400000)
Due to the compression of the cache key, the v2p cache incorrectly hits 0x0000000000400000 for 0x0000880000400000.